requirements.txt
streamlit
opencv-python
matplotlib
numpy
Pillow
